Error handling in separate file
