+++ src/Cody.VisualStudio/Cody.VisualStudio.csproj
@@ -53,6 +53,7 @@
     <Compile Include="Client\IAgentConnector.cs" />
     <Compile Include="Client\NameTransformer.cs" />
     <Compile Include="Client\RemoteAgentConnector.cs" />
+    <Compile Include="CodyPackage.ErrorHandling.cs" />
     <Compile Include="CodyToolWindow.cs" />
     <Compile Include="Infrastructure\LoggerFactory.cs" />
     <Compile Include="Infrastructure\WindowPaneLogger.cs" />
@@ -116,6 +117,9 @@
     <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="17.6.2164">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+    <PackageReference Include="Sentry">
+      <Version>4.13.0</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
