+++ src/Cody.VisualStudio/CodyPackage.cs
@@ -8,38 +8,31 @@ using Cody.Core.Logging;
 using Cody.Core.Settings;
 using Cody.Core.Workspace;
 using Cody.UI.Controls;
+using Cody.UI.ViewModels;
 using Cody.UI.Views;
 using Cody.VisualStudio.Client;
 using Cody.VisualStudio.Inf;
 using Cody.VisualStudio.Options;
 using Cody.VisualStudio.Services;
+using EnvDTE;
+using EnvDTE80;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Connected.CredentialStorage;
 using Microsoft.VisualStudio.Shell.Interop;
+using Microsoft.VisualStudio.TaskStatusCenter;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.Design;
 using System.IO;
-using System.Linq;
+using System.Net;
 using System.Reflection;
 using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Threading;
-using Cody.UI.ViewModels;
-using EnvDTE;
-using EnvDTE80;
-using Task = System.Threading.Tasks.Task;
-using Microsoft.VisualStudio.TaskStatusCenter;
 using SolutionEvents = Microsoft.VisualStudio.Shell.Events.SolutionEvents;
-using System.Net;
-using Newtonsoft.Json;
-using Cody.Core.Trace;
+using Task = System.Threading.Tasks.Task;
 
 namespace Cody.VisualStudio
 {
@@ -66,7 +59,7 @@ namespace Cody.VisualStudio
     [ProvideOptionPage(typeof(GeneralOptionsPage), "Cody", "General", 0, 0, true)]
     [ProvideToolWindow(typeof(CodyToolWindow), Style = VsDockStyle.Tabbed, Window = VsConstants.VsWindowKindSolutionExplorer)]
     [ProvideAutoLoad(VSConstants.UICONTEXT.SolutionOpening_string, PackageAutoLoadFlags.BackgroundLoad)]
-    public sealed class CodyPackage : AsyncPackage
+    public sealed partial class CodyPackage : AsyncPackage
     {
 
         public const string PackageGuidString = "9b8925e1-803e-43d9-8f43-c4a4f35b4325";
@@ -267,15 +260,6 @@ namespace Cody.VisualStudio
             }
         }
 
-
-        private void InitializeErrorHandling()
-        {
-            AppDomain.CurrentDomain.UnhandledException += CurrentDomainOnUnhandledException;
-            Application.Current.DispatcherUnhandledException += CurrentOnDispatcherUnhandledException;
-            TaskScheduler.UnobservedTaskException += TaskSchedulerOnUnobservedTaskException;
-        }
-
-
         private void PrepareAgentConfiguration()
         {
             try
@@ -438,62 +422,5 @@ namespace Cody.VisualStudio
             }
         }
 
-
-        private void CurrentDomainOnUnhandledException(object sender, UnhandledExceptionEventArgs e)
-        {
-            Logger.Error($"Unhandled domain exception:{e.ExceptionObject}");
-            Logger.Error($"Unhandled domain exception, is terminating:{e.IsTerminating}");
-
-            var exception = e.ExceptionObject as Exception;
-            Logger.Error("Unhandled domain exception occurred.", exception);
-
-        }
-
-        private void CurrentOnDispatcherUnhandledException(object sender, DispatcherUnhandledExceptionEventArgs e)
-        {
-            var exception = e.Exception;
-            Logger.Error("Unhandled exception occurred on the UI thread.", exception);
-
-            if (!System.Diagnostics.Debugger.IsAttached)
-                e.Handled = true;
-        }
-
-        private void TaskSchedulerOnUnobservedTaskException(object sender, UnobservedTaskExceptionEventArgs e)
-        {
-            try
-            {
-                if (!System.Diagnostics.Debugger.IsAttached)
-                    e.SetObserved();
-
-                var thirdPartyException = e.Exception;
-                var exceptionDetails = new StringBuilder();
-                if (thirdPartyException != null)
-                {
-                    exceptionDetails.AppendLine(thirdPartyException.Message);
-                    exceptionDetails.AppendLine(thirdPartyException.StackTrace);
-
-                    if (thirdPartyException.InnerExceptions.Any())
-                    {
-                        foreach (var inner in thirdPartyException.InnerExceptions)
-                        {
-                            exceptionDetails.AppendLine(inner.Message);
-                            exceptionDetails.AppendLine(inner.StackTrace);
-                        }
-                    }
-
-                    if (!exceptionDetails.ToString().Contains("Cody")) return;
-                }
-
-                Logger.Error("Unhandled exception occurred on the non-UI thread.", e.Exception);
-                foreach (var ex in e.Exception.InnerExceptions)
-                {
-                    Logger.Error("Inner exception", ex);
-                }
-            }
-            catch
-            {
-                // catching everything because if not VS will freeze/crash on the exception
-            }
-        }
     }
 }
